requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
backcall==0.1.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
discord.py==1.3.3
idna==2.9
IMDbPY==6.8
ipykernel==5.3.0
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
jupyter-client==6.1.3
jupyter-core==4.6.3
lxml==4.5.1
multidict==4.7.6
parso==0.7.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
Pygments==2.6.1
python-dateutil==2.8.1
pyzmq==19.0.1
requests==2.23.0
six==1.15.0
SQLAlchemy==1.3.17
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
websockets==8.1
yarl==1.4.2
